Intellij remarks (#129)
* make final where possible

* More Nullabilities

* Removing unnecessary throws.

* Unused imports

* Method is equal to its super.

* rounding with ints

* More unused imports

* Simplify boolean expressions.

* Remove redundant initializer

* Remove unused assignment

* cast to long

* unreachable code

* Pointless arithmetic expression.

* Add type cast

* Unary plus

* integer division in floating operation

* Use stringbuilder

* Suppress super call

* More finals

* remove toStrings

* remove modifiers

* remove semicolons

* replace switch with if

* inline variables

* use Objects.equals

* Remove empty if/else branches

* Deleting commented out code.

* Unnecessary import

* javadoc empty lines

* Remove dangling javadoc comments

* wrong parameters in javadocs

* Removing unused params in javadocs

* replacing links with real links

* difference in return type

* Inner classes may be static

* x * (int) 0.5 == x * 0 == 0

* Adding Nullability

* Unused replaceWhitelist

* Suspicious indentation

* assigning x/y/z from nr/chunkx/chunkZ is weird

* Convert to local fields

* return value isn't used

* Inline default values

* make it an int

* Safe delete unused methods

* editorconfig cleanup

* redundant suppressions

* fix building redirect

* Remove empty methods

* Beta-api

* make unused variables local

* was always true

* inline variables

* unused imports

* small cleanup

+++ common/src/main/java/org/terraform/structure/pillager/mansion/MansionCompoundRoomDistributor.java
@@ -63,9 +63,9 @@ public class MansionCompoundRoomDistributor {
 		}
 		potentialRoomSizes.add(new MansionRoomSize(2,2)); //At least one 2x2 room
 		
-		while(occupiedCells/pieces.size() < 0.7 || 
+		while((double) occupiedCells /pieces.size() < 0.7 ||
 				GenUtils.chance(random, pieces.size()-occupiedCells/4, pieces.size())) {
-			if(occupiedCells/pieces.size() < 0.5 && GenUtils.chance(random, 1, 3)) {
+			if((double) occupiedCells /pieces.size() < 0.5 && GenUtils.chance(random, 1, 3)) {
 				occupiedCells += 4;
 				potentialRoomSizes.add(new MansionRoomSize(2,2));
 			} else {
@@ -118,11 +118,8 @@ public class MansionCompoundRoomDistributor {
 	/**
 	 * Also sets the needed rooms to empty room populator and knocks down relevant walls
 	 * if the return value is true.
-	 * @param piece
 	 * @param pieces NOT TO BE MODIFIED
-	 * @param roomSize
-	 * @return
-	 */
+     */
 	public static boolean canRoomSizeFitWithCenter(@NotNull MansionStandardRoomPiece piece, @NotNull Collection<JigsawStructurePiece> pieces, @NotNull MansionRoomSize roomSize, @NotNull MansionRoomPopulator defaultPopulator, boolean force) {
 		
 		SimpleLocation center = piece.getRoom().getSimpleLocation();
